Adding in Audio Sources
Walking, Click for continue buttons.

Character controller cause I commented out option for audio but it didn't work. Added new script - walking sounds to enable based on if WASD was being pressed instead.
Updated continue button and all other UI buttons with click sound.
Increased speed of animation for walking to match a bit better with walking sound. It's not perfect but it's there for now.